requirements.txt
streamlit
pillow
easyocr
dateparser
streamlit-paste-button
numpy
icalendar